moved keyboard instantiation to the baseTool
